% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R, R/mg_create_template_raster.R
\name{mg_func_template_raster}
\alias{mg_func_template_raster}
\title{mg_create_template_raster}
\usage{
mg_func_template_raster(df_col_x, df_col_y, resolution, crs_to_assign)

mg_func_template_raster(df_col_x, df_col_y, resolution, crs_to_assign)
}
\arguments{
\item{df_col_x}{The column of the dataset with the x coordinate (should be numeric)}

\item{df_col_y}{The column of the dataset with the x coordinate (should be numeric)}

\item{resolution}{Resolution of the raster which is created (an integer, eg. 100 for 100x100 raster cell size)}

\item{crs_to_assign}{Coordinate reference system to assign to the created raster (integer, eg. 25832 for UTM 32, 31468 for Gauss krueger)}
}
\value{
a list with two elements, first is a grid, second is a raster file

a list with two elements, first is a grid, second is a raster file
}
\description{
Takes a data set with coordinates and creates a template raster with a specifies resolution and crs

Takes a data set with coordinates and creates a template raster with a specifies resolution and crs
}
\examples{
grid_project_area <- mg_func_template_raster(df_col_x = meta_projekt$ostwert_crs_25832,
df_col_y = meta_projekt$nordwert_crs_25832,
resolution = 200,
crs_to_assign = 25832)

grid_project_area <- mg_func_template_raster(df_col_x = meta_projekt$ostwert_crs_25832,
df_col_y = meta_projekt$nordwert_crs_25832,
resolution = 200,
crs_to_assign = 25832)

}
