% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mg_function_extract_spreaded_ts.R
\name{mg_function_extract_spreaded_ts}
\alias{mg_function_extract_spreaded_ts}
\title{mg_function_extract_spreaded_ts}
\usage{
mg_function_extract_spreaded_ts(start_zeitpunkt, input_settings, df)
}
\arguments{
\item{start_zeitpunkt}{starting time for which the time series should be extracted. Has to be same as from function mg_function_check_number_complete_time_series}

\item{input_settings}{dataframe with the input settings of function mg_function_check_number_complete_time_series, generated by this function}

\item{df}{dataframe which contains the value and date column}
}
\value{
dataframe of the time series in wide format with equal distances:
}
\description{
extract a spreaded time series from the dataframe used in function mg_function_check_number_complete_time_series
}
\examples{
spreaded_ts<- mg_function_extract_spreaded_ts(start_zeitpunkt = df_complete_time_series$number_of_timeseries$start_zeitpunkt[1],
                                              input_settings = df_complete_time_series$input_settings,
                                              df = wasserstand_projekt)

}
