% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mg_function_impute_spreaded_ts.R
\name{mg_function_impute_spreaded_ts}
\alias{mg_function_impute_spreaded_ts}
\title{mg_function_impute_spreaded_ts}
\usage{
mg_function_impute_spreaded_ts(
  spreaded_ts,
  plot_impute,
  export_path = NA,
  imputeTS_model = "StructTS",
  imputeTS_type = "trend"
)
}
\arguments{
\item{spreaded_ts}{spreaded df with time series data (e.g. from function mg_function_extract_spreaded_ts)}

\item{plot_impute}{True/False if time series with imputed values should be plotted}

\item{export_path}{export path for plots}

\item{imputeTS_model}{type of model passed to impute TS function, default is StructTS}

\item{imputeTS_type}{type of type passed to impute TS function, default is "trend"}
}
\value{
dataframe of the time series in wide format with equal distances:
}
\description{
impute spreaded time series using package imputeTS
}
\examples{
spreaded_ts<- mg_function_extract_spreaded_ts(start_zeitpunkt = df_complete_time_series$number_of_timeseries$start_zeitpunkt[1],
                                              input_settings = df_complete_time_series$input_settings,
                                              df = wasserstand_projekt)

}
