% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mg_function_longest_complete_ts.R
\name{mg_function_longest_complete_ts}
\alias{mg_function_longest_complete_ts}
\title{mg_function_longest_complete_ts}
\usage{
mg_function_longest_complete_ts(
  df,
  date_column_name,
  value_column_name,
  id_column_name,
  n_allowed_missing_values_days,
  minimum_timespan_ts_years
)
}
\arguments{
\item{df}{dataframe which contains the value and date column}

\item{date_column_name}{column name of the date column (should be format date)}

\item{value_column_name}{column name of the value column (numeric format)}

\item{id_column_name}{column name ofidentifier column (e.g. name of Grundwasssermessstelle) (character format)}

\item{n_allowed_missing_values_days}{timespan in number of days for which missing values are allowed}

\item{minimum_timespan_ts_years}{number of years which the ts should have at least (float)}
}
\value{
subset of initial dataframe which satisfies the conditions minimum_timespan_ts_years and n_allowed_missing_values_days
}
\description{
Takes hydraulic head time series dataframe and extracts for each dataset the longest complete time series
}
\examples{
mg_function_longest_complete_ts(df = wasserstand_projekt,
                               date_column_name = "datum",
                               value_column_name = "mess_wert",
                               id_column_name = datenquelle,
                               n_allowed_missing_values_days = 60,
                               minimum_timespan_ts_years = 2.5)
}
