% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mg_time_series_prep.R
\name{mg_function_time_series_preparation}
\alias{mg_function_time_series_preparation}
\title{mg_time_series_prep}
\usage{
mg_function_time_series_preparation(
  df,
  date_column_name,
  value_column_name,
  timespan_weeks
)
}
\arguments{
\item{df}{dataframe which contains the value and date column}

\item{date_column_name}{column name of the date column (should be format date)}

\item{value_column_name}{column name of the value column (numeric format)}

\item{timespan_weeks}{number of weeks for which mean value and mean date are calculated}
}
\value{
input dataframe with 4 new columns,
timespan_weeks: mean isoweek of timespan (e.g. isoweek 49-51 --> 50)
timespan_weeks_label: character column showing the isoweeks e.g. '49-51'
mean_value: mean value in time span
mean_date: mean date in time span (format date)
}
\description{
Takes hydraulic head time series datafram and calculates mean values and dates for a certain timespan.
df,date_column_name,value_column_name,timespan
}
\examples{
test<- mg_function_time_series_preparation(df = wasserstand_projekt,
date_column_name = "datum",
value_column_name = "mess_wert",
timespan = 3)

}
